requirements.txt
docopt==0.6.2
numpy==2.1.1
paho-mqtt==2.1.0
passlib==1.7.4
psutil==6.0.0
PyYAML==6.0.2
six==1.16.0
tenseal==0.3.15
transitions==0.9.2
websockets==13.1
